+++ gazon-microservice-starter/src/main/java/com/gitlab/exception/handler/WrongSelectedProductsException.java
@@ -1,0 +1,10 @@
+package com.gitlab.exception.handler;
+
+import org.springframework.http.HttpStatus;
+
+public class WrongSelectedProductsException extends BusinessException {
+
+    public WrongSelectedProductsException(HttpStatus httpStatus, String message) {
+        super(httpStatus, message);
+    }
+}